requirement.txt
langchain-openai
langchain
langgraph
PyPDF2
python-dotenv
requests